vinyl: vy_read_iterator_cmp explicitly says to move front
According to #2129 it is possible, that there is REPLACE and
DELETE with the same lsn and the same key. In such a case the DELETE
must be returned and REPLACE skipped, so vy_read_iterator_cmp must say,
the that DELETE precedes the REPLACE (sometimes by returning 1).
But in any case front it must be moved, becase actually it is
two statements with the same key. It is not possible by moving front
id on vy_read_iterator_cmp <= 0.